+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Rename useMemoSafe to useMemoPrev
 
 ## [1.1.0] - 2019-02-18
 ### Changed